+++ web/test/tests/segmentation-test.js
@@ -1,8 +1,5 @@
 'use strict';
 
-module.exports = function() {};
-
-
 var lib = require('../test-lib.js');
 
 module.exports = function(fw) {