+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Please add changes to "master", preferably ordered by their significance. (Most 
  - SQL boxes can optionally persist their outputs.
  - _Train a decision tree classification model_ can use String attributes.
  - Navigate directories by clicking on the path segments on the splash page.
+ - The new box metadata editor makes it possible to change the type of a box and fix broken custom
+   box references.
  - Popups can be more flexibly resized.
  - Arrows can be deleted by clicking on them.
  - A number of minor bugfixes and improvements. Thanks for the feedback!